+++ packages/extension-playground/src/App.tsx
@@ -24,7 +24,7 @@
 
  */
 import React from 'react'
-import { ExtensionProvider } from '@looker/extension-sdk-react'
+import { ExtensionProvider } from 'looker-extension-sdk-react'
 import { hot } from 'react-hot-loader/root'
 
 import { Playground } from './Playground'