+++ packages/hackathon/src/data/admin/sagas.ts
@@ -26,8 +26,8 @@
 import { all, call, put, takeEvery } from 'redux-saga/effects'
 import { SagaIterator } from 'redux-saga'
 import { ValidationMessages } from '@looker/components'
-import { getCore40SDK } from '@looker/extension-sdk-react'
-import { getExtensionSDK } from '@looker/extension-sdk'
+import { getCore40SDK } from 'looker-extension-sdk-react'
+import { getExtensionSDK } from 'looker-extension-sdk'
 import { IUserAttribute } from '@looker/sdk'
 import { actionMessage, beginLoading, endLoading } from '../common/actions'
 import {